fix: check of type context raising invalid errors
