+++ packages/safe-ds-lang/src/language/validation/other/types/unionTypes.ts
@@ -1,62 +1,12 @@
-import {
-    isSdsAnnotation,
-    isSdsCallable,
-    isSdsClass,
-    isSdsFunction,
-    isSdsParameter,
-    isSdsUnionType,
-    SdsUnionType,
-} from '../../../generated/ast.js';
-import { AstNode, AstUtils, ValidationAcceptor } from 'langium';
+import { SdsUnionType } from '../../../generated/ast.js';
+import { ValidationAcceptor } from 'langium';
 import { getTypeArguments } from '../../../helpers/nodeProperties.js';
 import { SafeDsServices } from '../../../safe-ds-module.js';
 import { Type } from '../../../typing/model.js';
 import { isEmpty } from '../../../../helpers/collections.js';
-import { getNodesToCheckForContextProvider } from './getNodesToCheckForContext.js';
 
-export const CODE_UNION_TYPE_CONTEXT = 'union-type/context';
 export const CODE_UNION_TYPE_DUPLICATE_TYPE = 'union-type/duplicate-type';
 export const CODE_UNION_TYPE_MISSING_TYPES = 'union-type/missing-types';
-
-export const unionTypeMustBeUsedInCorrectContext = (services: SafeDsServices) => {
-    const getNodesToCheckForContext = getNodesToCheckForContextProvider(services);
-
-    return (node: SdsUnionType, accept: ValidationAcceptor): void => {
-        const nodesToCheck = getNodesToCheckForContext(node);
-
-        for (const nodeToCheck of nodesToCheck) {
-            if (!contextIsCorrect(nodeToCheck)) {
-                accept(
-                    'error',
-                    'Union types must only be used for parameters of annotations, classes, and functions.',
-                    {
-                        node: nodeToCheck,
-                        code: CODE_UNION_TYPE_CONTEXT,
-                    },
-                );
-            }
-        }
-    };
-};
-
-const contextIsCorrect = (node: AstNode): boolean => {
-    if (AstUtils.hasContainerOfType(node.$container, isSdsUnionType)) {
-        return true;
-    }
-
-    const container = node.$container;
-    if (!isSdsParameter(container)) {
-        return false;
-    }
-
-    const containingCallable = AstUtils.getContainerOfType(container, isSdsCallable);
-    return (
-        !containingCallable ||
-        isSdsAnnotation(containingCallable) ||
-        isSdsClass(containingCallable) ||
-        isSdsFunction(containingCallable)
-    );
-};
 
 export const unionTypeMustHaveTypes = (node: SdsUnionType, accept: ValidationAcceptor): void => {
     if (isEmpty(getTypeArguments(node.typeArgumentList))) {
