+++ packages/safe-ds-lang/tests/resources/validation/other/types/union types/context/nested.sdsdev
@@ -6,7 +6,7 @@ package tests.validation.other.types.unionTypes.context
 
 class MyClass1 {
     // $TEST$ no error "Union types must only be used for parameters of annotations, classes, and functions."
-    attr a1: union<Int, »union<Int>«>
+    attr a1: union<Int, »union<Int, String>«>
     // $TEST$ no error "Union types must only be used for parameters of annotations, classes, and functions."
     attr a2: union<Int, »Alias«>
 }
@@ -14,7 +14,7 @@ class MyClass1 {
 class MyClass2 {
     // $TEST$ no error "Union types must only be used for parameters of annotations, classes, and functions."
     // $TEST$ no error "Union types must only be used for parameters of annotations, classes, and functions."
-    attr a1: union<Int, (p: »union<Int>«) -> (r: »union<Int>«)>
+    attr a1: union<Int, (p: »union<Int, String>«) -> (r: »union<Int, String>«)>
     // $TEST$ no error "Union types must only be used for parameters of annotations, classes, and functions."
     // $TEST$ no error "Union types must only be used for parameters of annotations, classes, and functions."
     attr a2: union<Int, (p: »Alias«) -> (r: »Alias«)>
